Welp. Doing some testing means that you can actually install the database now. It also re-writes the file to remove all of the installer stuff once it's done. Talk about destructive!
Codger - \cod-JA\ - noun; An old person.

Ship's log. Day 44.
Another man was reported missing today. That makes eight this week. Doctor Hanley has pointed out the fact that no-one can remember where each of the missing crew slept, despite being adamant that the person is no longer with us on board. Doctor Hanley suggests that this may in fact be an extention of what he has come to call 'The Effect'. The ship's engines were all out for a full day yesterday, however all instruments report that we neither slowed nor changed course in that time. Two officers, Harris and Pellaro, had to be confined to their quarters due to an incident related to The Effect. Morale is getting low.