Added a home page, text summarizer, and markdown code. Also updated the ReadMe with the changes to the MarkDown syntax. Wooooo.
Dunny - \dun-EE\ - noun; Toilet, most commonly an outside toilet. See: outhouse.

Ship's log. Day 55.
Half the crew are missing or dead. The Effect seems not to be passive, but intelligent and malevolent. Rare moments of clarity are more of a torture than a boon, as we realise how dire the situation has become. Rarely is more than one person unaffected at a time, so you are left to question a reality which no-one else seems to share. Doctor Hanley has suggested that perhaps these crew members who are missing (if, in fact they ever existed at all) may simply be held in a different reality. Hanley and I have managed to communicate through a notepad in the galley, which so far seems to be largely unaffected. I have not seen the man in person for several days. I hope that the messages I send are truly being recieved, and the responses I get are real. If not, all hope is lost.